requirements.txt
django
djangorestframework
django-cors-headers
pillow
channels
daphne
channels-redis
